Fix: 디토 삭제 s3 에러 수정

+++ src/main/java/site/dittotrip/ditto_trip/ditto/service/DittoService.java
@@ -86,8 +86,10 @@ public class DittoService {
         }
 
         // image 처리
-        String imagePath = s3Service.uploadFile(multipartFile);
-        ditto.setImagePath(imagePath);
+        if(!multipartFile.isEmpty()) {
+            String imagePath = s3Service.uploadFile(multipartFile);
+            ditto.setImagePath(imagePath);
+        }
 
         // 알림 처리
 //        processAlarmInSaveDitto(ditto);
@@ -129,7 +131,9 @@ public class DittoService {
             throw new NoAuthorityException();
         }
 
-        s3Service.deleteFile(ditto.getImagePath());
+        if(ditto.getImagePath() != null){
+            s3Service.deleteFile(ditto.getImagePath());
+        }
         dittoRepository.delete(ditto);
     }
